requirements.txt
alembic==1.11.1
blinker==1.8.2
click==8.1.7
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
greenlet==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.19.0
packaging==24.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
SQLAlchemy==2.0.34
typing_extensions==4.12.2
Werkzeug==3.0.4
